First Part Complete
This commit marks the end of the first part. All the documentation for
the code is being done within itself, so no documentation will be
written in here.

So far we wrote basic gameplay aspects, controls, sprites and added 3
more options for player sprites.

Next step will be to refactor the code and make it more readable and friendly.