Add preload back to game.js
For development and debugging purposes we might be skipping some of the
stages of the game, to avoid any errors being raised. We preload all the
sprites inside the game.js file.

This preload function is currently commented out and should only be used
when the game.stage.start argument changes to 'Game'.